+++ spec/lib/dawn/codesake_knowledgebase_spec.rb
@@ -898,4 +898,9 @@ end
         sc.should_not   be_nil
           sc.class.should == Codesake::Dawn::Kb::CVE_2014_0130
   end
+  it "must have test for CVE-2013-2105" do
+    sc = kb.find("CVE-2013-2105")
+    sc.should_not   be_nil
+    sc.class.should == Codesake::Dawn::Kb::CVE_2013_2105
+  end
 end