+++ spec/lib/kb/cve_2013_2105_spec.rb
@@ -1,0 +1,11 @@
+require 'spec_helper'
+describe "The CVE-2013-2105 vulnerability" do
+	before(:all) do
+		@check = Codesake::Dawn::Kb::CVE_2013_2105.new
+		# @check.debug = true
+	end
+  it "is reported when show_in_browser vulnerable version is reported (0.0.3)" do
+    @check.dependencies = [{:name=>'show_in_browser', :version=>'0.0.3'}]
+    @check.vuln?.should   be_true
+  end
+end