+++ lib/codesake/dawn/knowledge_base.rb
@@ -173,6 +173,7 @@ require "codesake/dawn/kb/cve_2013_1947"
 require "codesake/dawn/kb/cve_2013_1948"
 require "codesake/dawn/kb/cve_2013_2065"
 require "codesake/dawn/kb/cve_2013_2090"
+require "codesake/dawn/kb/cve_2013_2105"
 require "codesake/dawn/kb/cve_2013_2119"
 require "codesake/dawn/kb/cve_2013_2512"
 require "codesake/dawn/kb/cve_2013_2513"
@@ -420,6 +421,7 @@ module Codesake
           Codesake::Dawn::Kb::CVE_2013_1948.new,
           Codesake::Dawn::Kb::CVE_2013_2065.new,
           Codesake::Dawn::Kb::CVE_2013_2090.new,
+          Codesake::Dawn::Kb::CVE_2013_2105.new,
           Codesake::Dawn::Kb::CVE_2013_2119.new,
           Codesake::Dawn::Kb::CVE_2013_2512.new,
           Codesake::Dawn::Kb::CVE_2013_2513.new,