+++ lib/codesake/dawn/knowledge_base.rb
@@ -223,6 +223,7 @@ require "codesake/dawn/kb/cve_2014_2538"
 
 require "codesake/dawn/kb/osvdb_105971"
 require "codesake/dawn/kb/osvdb_108569"
+require "codesake/dawn/kb/osvdb_108570"
 
 
 module Codesake
@@ -468,7 +469,7 @@ module Codesake
           # OSVDB Checks are still here since are all about dependencies
           Codesake::Dawn::Kb::OSVDB_105971.new,
           Codesake::Dawn::Kb::OSVDB_108569.new,
-
+          Codesake::Dawn::Kb::OSVDB_108570.new,
         ]
         # END @cve_security_checks array
         # START @owasp_ror_cheatsheet_checks array