+++ spec/lib/dawn/codesake_knowledgebase_spec.rb
@@ -908,4 +908,10 @@ describe "The Codesake Dawn knowledge base" do
     sc.should_not   be_nil
     sc.class.should == Codesake::Dawn::Kb::OSVDB_108569
   end
+
+  it "must have test for OSVDB-108570" do
+      sc = kb.find("OSVDB-108570")
+        sc.should_not   be_nil
+          sc.class.should == Codesake::Dawn::Kb::OSVDB_108570
+  end
 end