+++ spec/lib/dawn/codesake_knowledgebase_spec.rb
@@ -920,4 +920,9 @@ describe "The Codesake Dawn knowledge base" do
         sc.should_not   be_nil
           sc.class.should == Codesake::Dawn::Kb::OSVDB_108530
   end
+  it "must have test for OSVDB-108563" do
+      sc = kb.find("OSVDB-108563")
+        sc.should_not   be_nil
+          sc.class.should == Codesake::Dawn::Kb::OSVDB_108563
+  end
 end