+++ spec/lib/kb/cve_2014_3483_spec.rb
@@ -1,0 +1,24 @@
+require 'spec_helper'
+describe "The CVE-2014-3483 vulnerability" do
+	before(:all) do
+		@check = Codesake::Dawn::Kb::CVE_2014_3483.new
+		# @check.debug = true
+	end
+  it "must be filled with CVSS information"
+  it "is reported when a rails gem version 4.0.6 is detected" do
+    @check.dependencies = [{:name=>"rails", :version=>"4.0.6"}]
+    @check.vuln?.should   be_true
+  end
+  it "is reported when a rails gem version 4.1.2 is detected" do
+    @check.dependencies = [{:name=>"rails", :version=>"4.1.2"}]
+    @check.vuln?.should   be_true
+  end
+  it "is not reported when a rails gem version 4.0.7 is detected" do
+    @check.dependencies = [{:name=>"rails", :version=>"4.0.7"}]
+    @check.vuln?.should   be_false
+  end
+  it "is not reported when a rails gem version 4.1.3 is detected" do
+    @check.dependencies = [{:name=>"rails", :version=>"4.1.3"}]
+    @check.vuln?.should   be_false
+  end
+end