+++ Changelog.md
@@ -7,6 +7,8 @@ frameworks.
 
 _latest update: Fri Apr 18 07:55:10 CEST 2014_
 
+* Adding a check for CVE-2013-2105: HTML injection in show_in_browser rubygem
+
 ## Version 1.1.3 - codename: Lightning McQueen (2014-05-06)
 
 * Adding a check for CVE-2014-0130: directory traversal for ruby on rails