+++ Changelog.md
@@ -11,12 +11,17 @@ _latest update: Fri Apr 18 07:55:10 CEST 2014_
   gem (issue #69)
 * Fix issue #62 about codesake-dawn config filename
 * Adding a check for CVE-2013-2105: HTML injection in show_in_browser rubygem
+<<<<<<< HEAD
 * Fixing issue #68. I moved some file variable operation after the check for
   nil value. I can't reproduce on my station the offending bug :-(
 * Adding a check for OSVDB-108570: arbitrary command execution for
   backup_checksum gem (issue #70)
-* Adding a check for OSVDB-108530: remote code execution for kajam gem (issue #71)
-* Adding a check for OSVDB-108563: remote code execution in gyazo rubygem (issue #72)
+* Adding a check for OSVDB-108530: remote code execution for kajam gem (issue
+  #71)
+* Adding a check for OSVDB-108563: remote code execution in gyazo rubygem
+  (issue #72)
+* Adding a check for CVE-2014-3482: SQL Injection on Rails 3.2.x due to
+  PostgreSQL adapter for active record (issue #73)
 
 ## Version 1.1.3 - codename: Lightning McQueen (2014-05-06)